add frame skip control mechanism

+++ My/train_13.py
@@ -27,6 +27,8 @@ def get_reward(base_model, env, ep_max_step, sigma, CONFIG, seed_and_id=None):
         model = base_model
     env.frameskip = 1
     observation = env.reset()
+    obs_list = [observation] * FRAME_SKIP
+    break_is_true = False
     ep_r = 0.
     # print('k_id mid:', k_id,time.time()-start)
     if ep_max_step is None:
@@ -35,17 +37,23 @@ def get_reward(base_model, env, ep_max_step, sigma, CONFIG, seed_and_id=None):
         for step in range(ep_max_step):
             # print(trans(observation).size())
             # print(type(observation))
-            action = model(observation)[0].argmax().item()
-            # print(action)
-            observation, reward , done, _ = env.step(action)
-            ep_r += reward
-            if done:
+            if len(obs_list) != FRAME_SKIP:
+                print("Error", len(obs_list), "step:",step)
+            action = model(obs_list)[0].argmax().item()
+            obs_list = []
+            for i in range(FRAME_SKIP):
+                observation, reward , done, _ = env.step(action)
+                obs_list.append(observation)
+                ep_r += reward
+                if done:
+                    break_is_true = True
+            if break_is_true:
                 break
     # print('k_id final:', k_id,time.time()-start)
     # print('k_id step:', k_id,step)
     return ep_r, step
 
-def train(model, optimizer, utility, pool, sigma, env, N_KID, CONFIG):
+def train(model, optimizer, pool, sigma, env, N_KID, CONFIG):
     # pass seed instead whole noise matrix to parallel will save your time
     # mirrored sampling
     noise_seed = np.random.randint(0, 2 ** 32 - 1, size=N_KID, dtype=np.uint32).repeat(2)   
@@ -58,14 +66,20 @@ def train(model, optimizer, utility, pool, sigma, env, N_KID, CONFIG):
     rewards = []
     timesteps = []
     timesteps_count = 0
-    for j in jobs:
+    for idx, j in enumerate(jobs):
         rewards.append(j.get()[0])
         timesteps.append(j.get()[1])
         timesteps_count += j.get()[1]
-        if timesteps_count > timesteps_per_batch:
+        if timesteps_count > timesteps_per_batch and idx % 2 == 1:
+            # make sure 2*Kid_N offspring have been tested
             pool.terminate()
             break
     
+    base = len(rewards)
+    rank = np.arange(1, base + 1)
+    util_ = np.maximum(0, np.log(base / 2 + 1) - np.log(rank))
+    utility = util_ / util_.sum() - 1 / base
+
     kids_rank = np.argsort(rewards)[::-1]               # rank kid id by reward
 
     cumulative_update = {}       # initialize update values
